chore(deps): break BlkExportServiceCore <-> exporters cycle
Extracts IBlkExportResult and ExportableUnitState union into a new
BlkExportServiceCore.types.ts leaf. Both BlkExportServiceCore.ts and
BlkExportServiceCore.exporters.ts import from the leaf; the orchestrator
re-exports for backward compat with any external callers.

Pattern 1 mechanical fix (same as PRs #419, #420, #422). 5th cycle of
the 6 deferred Pattern 4 cycles closed.
